+++ admin/assets/javascripts/discourse/routes/admin-plugins/show/topic-reply-limits/usage.js
@@ -1,0 +1,15 @@
+import { ajax } from "discourse/lib/ajax";
+import DiscourseRoute from "discourse/routes/discourse";
+
+export default class AdminPluginsShowTopicReplyLimitsUsageRoute extends DiscourseRoute {
+  async model() {
+    return await ajax(
+      "/admin/plugins/discourse-topic-reply-limits/usage.json"
+    );
+  }
+
+  setupController(controller, model) {
+    super.setupController(controller, model);
+    controller.initializeReport(model);
+  }
+}